#i10000#: cast to avoid overload ambiguity between SvSlotElement::Seek(unsigned long) and SvSlotElement::Seek(void*), hope this doesn't break 64bit build again

+++ idl/source/objects/slot.cxx
@@ -2,9 +2,9 @@
  *
  *  $RCSfile: slot.cxx,v $
  *
- *  $Revision: 1.13 $
+ *  $Revision: 1.14 $
  *
- *  last change: $Author: rt $ $Date: 2004-06-17 13:30:09 $
+ *  last change: $Author: hr $ $Date: 2004-06-21 11:32:19 $
  *
  *  The Contents of this file are made available subject to the terms of
  *  either of the following licenses
@@ -1223,7 +1223,7 @@ void SvMetaSlot::Insert( SvSlotElementList& rList, const ByteString & rPrefix,
         pLinkedSlot = pFirstEnumSlot;
 
         // Slaves untereinander verketten
-        rList.Seek(0U);
+        rList.Seek((ULONG)0);
         xEnumSlot = pFirstEnumSlot;
         SvSlotElement *pEle;
         do
